+++ test/unit/src/http/invoke-http/http/_req-header-fmt-test.js
@@ -2,6 +2,10 @@ let test = require('tape')
 let { join } = require('path')
 let sut = join(process.cwd(), 'src', 'http', 'invoke-http', 'http', '_req-header-fmt')
 let requestHeaderFormatter = require(sut)
+let params = {
+  req: { socket: { localPort: 'foo' } },
+  ip: 'bar',
+}
 
 test('Set up env', t => {
   t.plan(1)
@@ -10,19 +14,19 @@ test('Set up env', t => {
 
 test('Basics', t => {
   t.plan(2)
-  let result = requestHeaderFormatter()
+  let result = requestHeaderFormatter({}, params)
   t.ok(result.headers, 'Got back headers')
   t.notOk(result.cookies, 'Did not get back cookies without headers')
 })
 
 test('Header mangling & cookies (HTTP)', t => {
-  t.plan(8)
+  t.plan(11)
   let reqHeaders = {
     authorization: 'whatev',
     cookie: 'foo=bar; fiz=buz',
     Foo: 'bar'
   }
-  let { headers, cookies } = requestHeaderFormatter(reqHeaders, true)
+  let { headers, cookies } = requestHeaderFormatter(reqHeaders, params)
   // Positive casing checks
   t.ok(headers.authorization, 'Got back lowcased authorization')
   t.ok(headers.foo, 'Got back lowcased foo')
@@ -34,6 +38,10 @@ test('Header mangling & cookies (HTTP)', t => {
   t.equal(cookies.length, 2, 'Got back two cookies')
   t.equal(cookies[0], 'foo=bar', 'Got back cookie one')
   t.equal(cookies[1], 'fiz=buz', 'Got back cookie two')
+  // Metadata
+  t.equal(headers['x-forwarded-for'], 'bar', 'Got back header: x-forwarded-for')
+  t.equal(headers['x-forwarded-port'], 'foo', 'Got back header: x-forwarded-port')
+  t.equal(headers['x-forwarded-proto'], 'http', 'Got back header: x-forwarded-proto')
 })
 
 test('Header drops (HTTP)', t => {
@@ -47,6 +55,6 @@ test('Header drops (HTTP)', t => {
     'transfer-encoding': 'whatev',
     'upgrade': 'whatev',
   }
-  let { headers } = requestHeaderFormatter(reqHeaders)
-  t.equal(Object.keys(headers).length, 1, 'Dropped appropriate headers')
+  let { headers } = requestHeaderFormatter(reqHeaders, params)
+  t.equal(Object.keys(headers).length, 4, 'Dropped appropriate headers')
 })