+++ test/unit/src/http/invoke-http/index-req-test.js
@@ -12,16 +12,6 @@ let { arc7, arc6, headers } = require('@architect/req-res-fixtures').http.req
 let inventory = { inv: { _project: { preferences: null } } }
 lambdaStub.yields(null, {})
 
-function apiGwHeaders (headers, v5) {
-  let normal = {}
-  Object.entries(headers).forEach(([ h, v ]) => {
-    let header = h.toLowerCase()
-    if (header === 'cookie' && v5) header = 'Cookie'
-    normal[header] = v
-  })
-  return normal
-}
-
 let url = i => `http://localhost:6666${i ? i : ''}`
 let str = i => JSON.stringify(i)
 let match = (copy, item) => `${copy} matches: ${str(item)}`
@@ -30,6 +20,7 @@ let response = {
   setHeader: sinon.fake.returns(),
   end: sinon.fake.returns()
 }
+let httpVersion = '1.1'
 
 function teardown () {
   lambdaStub.reset() // mostly jic
@@ -75,8 +66,8 @@ function checkArcV7HttpResult (mock, req, t) {
       if (timeEpoch) ref.timeEpoch = timeEpoch
     }
     t.equal(
-      str(ref),
       str(req[param]),
+      str(ref),
       match(`${param}`, req[param])
     )
   })
@@ -95,7 +86,8 @@ test('Architect v7 (HTTP API mode): get /', t => {
     url: url(), // Set by `router` (interpolated, API passes path param)
     body: {},   // {} set by `body-parser` (Arc 5 == {}, Arc 6 == null)
     headers,    // Set by requesting client
-    params: {}  // {} set by `router` (Arc 5 == {}, Arc 6 == null)
+    params: {}, // {} set by `router` (Arc 5 == {}, Arc 6 == null)
+    httpVersion,
   }
   handler(input, response)
   // Compare handler-generated request to mock
@@ -115,7 +107,8 @@ test('Architect v7 (HTTP API mode): get /?whats=up', t => {
     url: url('?whats=up'),
     body: {},
     headers,
-    params: {}
+    params: {},
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -134,7 +127,8 @@ test('Architect v7 (HTTP API mode): get /?whats=up&whats=there', t => {
     url: url('?whats=up&whats=there'),
     body: {},
     headers,
-    params: {}
+    params: {},
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -153,7 +147,8 @@ test('Architect v7 (HTTP API mode): get /nature/hiking', t => {
     url: url('/nature/hiking'),
     body: {},
     headers,
-    params: mock.pathParameters
+    params: mock.pathParameters,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -172,7 +167,8 @@ test('Architect v7 (HTTP API mode): get /{proxy+}', t => {
     url: url('/nature/hiking'),
     body: {},
     headers,
-    params: { '0': mock.pathParameters.proxy }
+    params: { '0': mock.pathParameters.proxy },
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -191,7 +187,8 @@ test('Architect v7 (HTTP API mode): get /path/* (/path/hi/there)', t => {
     url: url('/path/hi/there'),
     body: {},
     headers,
-    params: { '0': mock.pathParameters.proxy }
+    params: { '0': mock.pathParameters.proxy },
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -213,7 +210,8 @@ test('Architect v7 (HTTP API mode): get /:activities/{proxy+} (/nature/hiking/wi
     params: {
       activities: 'nature',
       '0': mock.pathParameters.proxy
-    }
+    },
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -233,7 +231,8 @@ test('Architect v7 (HTTP API mode): post /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: false // Assumes flag is set in binary handler
+    isBase64Encoded: false, // Assumes flag is set in binary handler
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -253,7 +252,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): post /form (form URL encoded)', 
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -273,7 +273,8 @@ test('Architect v7 (HTTP API mode): post /form (multipart form data)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true // Assumes flag is set in binary handler
+    isBase64Encoded: true, // Assumes flag is set in binary handler
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -293,7 +294,8 @@ test('Architect v7 (HTTP API mode): post /form (octet stream)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true // Assumes flag is set in binary handler
+    isBase64Encoded: true, // Assumes flag is set in binary handler
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -313,7 +315,8 @@ test('Architect v7 (HTTP API mode): put /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: false // Assumes flag is set in binary handler
+    isBase64Encoded: false, // Assumes flag is set in binary handler
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -333,7 +336,8 @@ test('Architect v7 (HTTP API mode): patch /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: false // Assumes flag is set in binary handler
+    isBase64Encoded: false, // Assumes flag is set in binary handler
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -353,7 +357,8 @@ test('Architect v7 (HTTP API mode): delete /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: false // Assumes flag is set in binary handler
+    isBase64Encoded: false, // Assumes flag is set in binary handler
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -366,7 +371,7 @@ test('Architect v7 (HTTP API mode): delete /form (JSON)', t => {
  */
 // Checks AWS's funky multiValueHeaders + multiValueQueryStringParameters
 function checkMultiValueHeaders (mock, req, t) {
-  let headers = apiGwHeaders(mock.headers)
+  let headers = mock.headers
   // Fixtures always have headers
   for (let header of Object.keys(headers)) {
     if (headers[header] !== req.multiValueHeaders[header][0])
@@ -398,8 +403,8 @@ function checkMultiValueQueryStringParameters (mock, req, t) {
 function checkArcV6RestResult (params, mock, req, t) {
   params.forEach(param => {
     t.equal(
-      str(mock[param]),
       str(req[param]),
+      str(mock[param]),
       match(`${param}`, req[param])
     )
   })
@@ -424,7 +429,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): get /', t => {
     url: url(), // Set by `router` (interpolated, API passes path param)
     body: {},   // {} set by `body-parser` (Arc 5 == {}, Arc 6 == null)
     headers,    // Set by requesting client
-    params: {}  // {} set by `router` (Arc 5 == {}, Arc 6 == null)
+    params: {}, // {} set by `router` (Arc 5 == {}, Arc 6 == null)
+    httpVersion,
   }
   handler(input, response)
   // Compare handler-generated request to mock
@@ -445,7 +451,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): get /?whats=up', t => {
     url: url('?whats=up'),
     body: {},
     headers,
-    params: {}
+    params: {},
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -465,7 +472,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): get /?whats=up&whats=there', t =
     url: url('?whats=up&whats=there'),
     body: {},
     headers,
-    params: {}
+    params: {},
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -485,7 +493,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): get /nature/hiking', t => {
     url: url('/nature/hiking'),
     body: {},
     headers,
-    params: mock.pathParameters
+    params: mock.pathParameters,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -505,7 +514,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): get /{proxy+}', t => {
     url: url('/nature/hiking'),
     body: {},
     headers,
-    params: { '0': mock.pathParameters.proxy }
+    params: { '0': mock.pathParameters.proxy },
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -525,7 +535,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): get /path/* (/path/hi/there)', t
     url: url('/path/hi/there'),
     body: {},
     headers,
-    params: { '0': mock.pathParameters.proxy }
+    params: { '0': mock.pathParameters.proxy },
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -548,7 +559,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): get /:activities/{proxy+} (/natu
     params: {
       activities: 'nature',
       '0': mock.pathParameters.proxy
-    }
+    },
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -569,7 +581,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): post /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true // Assumes flag is set in binary handler
+    isBase64Encoded: true, // Assumes flag is set in binary handler
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -590,7 +603,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): post /form (form URL encoded)', 
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -611,7 +625,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): post /form (multipart form data)
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -632,7 +647,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): post /form (octet stream)', t =>
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -653,7 +669,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): put /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -674,7 +691,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): patch /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -695,7 +713,8 @@ test('Architect v7 (HTTP + Lambda 1.0 payload): delete /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -719,7 +738,8 @@ test('Architect v6 (REST API mode): get /', t => {
     url: url(), // Set by `router` (interpolated, API passes path param)
     body: {},   // {} set by `body-parser` (Arc 5 == {}, Arc 6 == null)
     headers,    // Set by requesting client
-    params: {}  // {} set by `router` (Arc 5 == {}, Arc 6 == null)
+    params: {}, // {} set by `router` (Arc 5 == {}, Arc 6 == null)
+    httpVersion,
   }
   handler(input, response)
   // Compare handler-generated request to mock
@@ -740,7 +760,8 @@ test('Architect v6 (REST API mode): get /?whats=up', t => {
     url: url('?whats=up'),
     body: {},
     headers,
-    params: {}
+    params: {},
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -760,7 +781,8 @@ test('Architect v6 (REST API mode): get /?whats=up&whats=there', t => {
     url: url('?whats=up&whats=there'),
     body: {},
     headers,
-    params: {}
+    params: {},
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -780,7 +802,8 @@ test('Architect v6 (REST API mode): get /nature/hiking', t => {
     url: url('/nature/hiking'),
     body: {},
     headers,
-    params: mock.pathParameters
+    params: mock.pathParameters,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -802,7 +825,8 @@ test('Architect v6 (REST API mode): get /{proxy+}', t => {
     resource: '/{proxy+}', // The only time we should be using this
     body: {},
     headers,
-    params: mock.pathParameters // Would normally be { '0': ... }
+    params: mock.pathParameters, // Would normally be { '0': ... }
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -823,7 +847,8 @@ test('Architect v6 (REST API mode): post /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true // Assumes flag is set in binary handler
+    isBase64Encoded: true, // Assumes flag is set in binary handler
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -844,7 +869,8 @@ test('Architect v6 (REST API mode): post /form (form URL encoded)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -865,7 +891,8 @@ test('Architect v6 (REST API mode): post /form (multipart form data)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -886,7 +913,8 @@ test('Architect v6 (REST API mode): post /form (octet stream)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -907,7 +935,8 @@ test('Architect v6 (REST API mode): put /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -928,7 +957,8 @@ test('Architect v6 (REST API mode): patch /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event
@@ -949,7 +979,8 @@ test('Architect v6 (REST API mode): delete /form (JSON)', t => {
     body: mock.body,
     headers: mock.headers,
     params: {},
-    isBase64Encoded: true
+    isBase64Encoded: true,
+    httpVersion,
   }
   handler(input, response)
   let req = lambdaStub.args[0][0].event