+++ changelog.md
@@ -2,6 +2,19 @@
 
 ---
 
+## [5.7.0] 2023-06-27
+
+### Added
+
+- Added additional `@http` HTTP API request data
+  - `requestContext` properties: `http.protocol` + `http.sourceIp` + `http.userAgent`
+  - Headers: `user-agent` + `x-forwarded-for` + `x-forwarded-port` + `x-forwarded-proto`
+- Added additional `@http` REST API request data
+  - `requestContext` properties: `protocol` + `identity.sourceIp` + `identity.userAgent`
+  - Headers: `User-Agent` + `X-Forwarded-For` + `X-Forwarded-Port` + `X-Forwarded-Proto`
+
+---
+
 ## [5.6.10] 2023-06-21
 
 ### Fixed